State the intended integer type at the remaining mixed-width call sites
The same deduction failures as before, in the code that had not been reached yet: min, max and
div_rounding_up_safe given one int literal and one size_type, a lambda or ternary whose two
arms disagree once the types diverge, and a brace initializer that becomes narrowing. Nothing
here changes what is computed.

The cooperative-groups reduction in strings::slice is the one worth noting: it was summing
size_type character counts with a plus<int>, which cg::reduce rejects outright rather than
silently narrowing.

+++ cpp/src/io/json/nested_json_gpu.cu
@@ -1939,7 +1939,8 @@ void make_json_column(json_column& root_column,
 
       // Increment parent's child count and insert this struct node into the data path
       current_data_path.top().num_children++;
-      current_data_path.push({selected_col, target_row_index, nullptr, zero_child_count});
+      current_data_path.push(
+        {selected_col, static_cast<uint32_t>(target_row_index), nullptr, zero_child_count});
 
       // Add this struct node to the current column
       selected_col->append_row(target_row_index,
@@ -1977,7 +1978,8 @@ void make_json_column(json_column& root_column,
 
       // Increment parent's child count and insert this struct node into the data path
       current_data_path.top().num_children++;
-      current_data_path.push({selected_col, target_row_index, nullptr, zero_child_count});
+      current_data_path.push(
+        {selected_col, static_cast<uint32_t>(target_row_index), nullptr, zero_child_count});
 
       // Add this struct node to the current column
       selected_col->append_row(target_row_index,
